Add day 15: "Warehouse Woes", part 2.
This was a mess. I based part 1 on sets which worked well enough, but that turned out to be a massive headache for part 2. I ended up redoing the whole thing with lists instead and figured I'd merge it in with p1 when done, but I've had quite enough of this for now. Maybe some other time.