Add day 24 part 2.
The code is a mess because the problem isn't directly computable, you have to figure out which gates are connected the wrong way and the way I did it was to just check the output against what it's supposed to be based on the adder circuit. When something doesn't match up the way it's supposed to be, I just checked the puzzle input and figured out the wrong part manually.